add CircuitBuildTask and circuit unit tests
This commit removes the circuit FSM handshake and replaces it with
a CircuitBuildTask object. CircuitBuildTask is more flexible and
can be extended to support different types of circuits, and it also
removes the complexity of constructing a circuit from the Circuit
class itself. Circuits are now just "dumb" tunnels, merely forwarding
traffic to and from nodes, crypting cells, and handling flow control.

This commit also adds unit tests for the entire circuit package.