add weighted path algorithm and unit tests
This commit adds a full implementation of tor's real path seelction
algorithm, where the probability of selecting a node for a given position
is weighted both by the desired position, the node's bandwidth history, and
bandwidth parameters from the network consensus.

This implementation is *heavily* based on pathsim.py from the torps project.
The original version of the torps code can be found at:
https://github.com/torps/torps. Path selection in torps is basically a
straightforward port of the C path selection algorithm in tor itself.

The main modifications oppy makes are: making the code more modular and
testable, making some things more "pythonic", and providing deferred
wrappers so the code can be integrated into oppy.

Although based on torps, oppy and the modifications made to the torps code
are neither endorsed nor reviewed by the torps authors.

This commit also adds full unit tests for the path package.