add cellprocessor
Cell processing functions for circuits have been pulled out of the Circuit
class and now exist as standalone functions in cellprocessor that take a
circuit and a cell as an argument. This simplifies the Circuit class a bit
and it should, hopefully, make it easier to audit oppy for correct behavior
as it's now easier to pinpoint exactly what happens when a particular cell
is delivered to a circuit.